# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Veridian Lead Management API
  version: 1.0.0
  description: Private consultancy enquiry-to-onboarding API. Better Auth owns routes under /api/auth.
servers:
  - url: http://localhost:5000
paths:
...
  /api/v1/enquiries:
    post:
      summary: Submit a public enquiry
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Enquiry' } } } }
      responses:
        '202': { description: Durable generic receipt }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
...
  /api/v1/admin/leads:
    get:
      summary: Scoped paginated leads
      security: [ { sessionCookie: [] } ]
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: search, in: query, schema: { type: string, maxLength: 120 } }
        - { name: stage, in: query, schema: { $ref: '#/components/schemas/LeadStage' } }
      responses: { '200': { description: Scoped page }, '401': { $ref: '#/components/responses/Error' } }
  /api/v1/admin/leads/stats:
    get: { summary: Scoped pipeline counts with a date filter, security: [ { sessionCookie: [] } ], responses: { '200': { description: Counts } } }
  /api/v1/admin/leads/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get: { summary: Role-filtered lead detail, security: [ { sessionCookie: [] } ], responses: { '200': { description: Lead detail }, '404': { $ref: '#/components/responses/Error' } } }
    patch:
      summary: Update allowed lead fields with optimistic version
      security: [ { sessionCookie: [] } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [expectedVersion], properties: { expectedVersion: { type: integer }, fullName: { type: string }, email: { type: string, format: email }, country: { type: string }, assignedStaffId: { type: [string, 'null'] }, tierInterest: { $ref: '#/components/schemas/Tier' } }, additionalProperties: false } } } }
      responses: { '200': { description: Updated }, '409': { $ref: '#/components/responses/Error' } }
  /api/v1/admin/leads/{id}/notes:
    post: { summary: Add encrypted private note, security: [ { sessionCookie: [] } ], parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '201': { description: Created }, '403': { $ref: '#/components/responses/Error' } } }
  /api/v1/admin/leads/{id}/transitions:
    post: { summary: Perform an allowed manual transition, security: [ { sessionCookie: [] } ], parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: Updated }, '409': { $ref: '#/components/responses/Error' } } }
  /api/v1/admin/leads/{id}/archive:
    post: { summary: Archive and cancel discretionary queue, security: [ { sessionCookie: [] } ], parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: Archived } } }
  /api/v1/admin/leads/{id}/restore:
    post: { summary: Restore visibility without changing stage, security: [ { sessionCookie: [] } ], parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: Restored } } }
  /api/v1/admin/leads/{id}/communications:
    get: { summary: Scoped communication and attempt history, security: [ { sessionCookie: [] } ], parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: History } } }
  /api/v1/admin/enquiries:
    get: { summary: Reconciliation-pending enquiries, security: [ { sessionCookie: [] } ], responses: { '200': { description: Enquiries } } }
  /api/v1/admin/enquiries/{id}/reconcile:
    post: { summary: Link an enquiry without overwriting the lead, security: [ { sessionCookie: [] } ], parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: Reconciled }, '409': { $ref: '#/components/responses/Error' } } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: veridian.session_token }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 16, maxLength: 128 } }
  schemas:
...
    LeadStage: { type: string, enum: [NEW, CONTACTED, QUALIFICATION_PENDING, QUALIFIED, ENGAGEMENT_SENT, SIGNED, CONVERTED, REJECTED] }
    Tier: { type: string, enum: [FOUNDATION, ACCELERATED, EXECUTIVE, NOT_SURE] }
...
    Enquiry:
      type: object
      additionalProperties: false
      required: [fullName, email, country, netWorthRange, tierInterest, disclaimerAccepted, noticeVersion]
      properties:
        fullName: { type: string, minLength: 2, maxLength: 120 }
        email: { type: string, format: email, maxLength: 254 }
        country: { type: string, minLength: 2, maxLength: 100 }
        netWorthRange: { type: string, enum: ['$1M–$5M', '$1M-$5M', '$5M–$20M', '$5M-$20M', '$20M+', USD_1M_5M, USD_5M_20M, USD_20M_PLUS] }
        tierInterest: { type: string, enum: [Foundation, Accelerated, Executive, 'Not sure', FOUNDATION, ACCELERATED, EXECUTIVE, NOT_SURE] }
        disclaimerAccepted: { type: boolean, const: true }
        noticeVersion: { type: string }
        marketingConsent: { type: boolean, default: false }
        website: { type: string, maxLength: 0 }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            fieldErrors: { type: object, additionalProperties: { type: array, items: { type: string } } }
  responses:
    Error: { description: Safe error envelope, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }